backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
pymongo>=4.0.0
motor>=2.5.0  # MongoDB için asenkron sürücü
pika>=1.2.0
pandas>=1.3.3
python-dotenv>=0.19.1
websockets>=10.0
pytest>=6.2.5
httpx>=0.19.0
requests>=2.26.0
geopy>=2.2.0
numpy>=1.21.2 